# frontend/src/types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eval-studio",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/evaluations": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Create Evaluation",
        "description": "Create a new evaluation.",
        "operationId": "create_evaluation_api_v1_evaluations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "List Evaluations",
        "description": "List evaluations with pagination and optional filters.",
        "operationId": "list_evaluations_api_v1_evaluations_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EvaluationMode"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EvaluationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_EvaluationResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/run": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Run And Wait",
        "description": "Create an evaluation and run it, returning results in a single request.\n\nWhen ``async=true``, launches the evaluation as a background task and\nreturns 202 with a poll URL.  Otherwise blocks until the evaluation\ncompletes (or times out) and returns the full ``RunResponse``.\n\nContent negotiation: if the ``Accept`` header is ``text/plain``, the\nresponse body is ``{score}\\n{VERDICT}`` for easy CLI/pipeline usage.",
        "operationId": "run_and_wait_api_v1_evaluations_run_post",
        "parameters": [
          {
            "name": "async",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Async"
            }
          },
          {
            "name": "timeout",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Timeout"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Evaluation",
        "description": "Get an evaluation by ID.",
        "operationId": "get_evaluation_api_v1_evaluations__evaluation_id__get",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "evaluations"
        ],
        "summary": "Delete Evaluation",
        "description": "Delete an evaluation and all its results.",
        "operationId": "delete_evaluation_api_v1_evaluations__evaluation_id__delete",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}/cancel": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Cancel Evaluation",
        "description": "Cancel a running evaluation.",
        "operationId": "cancel_evaluation_api_v1_evaluations__evaluation_id__cancel_post",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}/run": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Run Evaluation",
        "description": "Trigger an evaluation run as a background task.",
        "operationId": "run_evaluation_api_v1_evaluations__evaluation_id__run_post",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/{evaluation_id}/rerun": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Rerun Evaluation",
        "description": "Re-run a completed or failed evaluation, clearing old results.",
        "operationId": "rerun_evaluation_api_v1_evaluations__evaluation_id__rerun_post",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvaluationCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable evaluation name."
          },
          "mode": {
            "$ref": "#/components/schemas/EvaluationMode",
            "description": "Evaluation mode: qa, rag, agent, or arena."
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id",
            "description": "ID of the dataset to evaluate against."
          },
          "judge_config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Config Id",
            "description": "ID of the judge configuration for scoring."
          },
          "rubric_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric Id",
            "description": "ID of the rubric to use for dimension-based scoring."
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config",
            "description": "Mode-specific configuration.",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "name",
          "mode"
        ],
        "title": "EvaluationCreate",
        "description": "Schema for creating an evaluation."
      },
      "EvaluationMode": {
        "type": "string",
        "enum": [
          "qa",
          "agent",
          "rag",
          "arena"
        ],
        "title": "EvaluationMode"
      },
      "EvaluationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique identifier for the evaluation."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable evaluation name."
          },
          "mode": {
            "$ref": "#/components/schemas/EvaluationMode",
            "description": "Evaluation mode: qa, rag, agent, or arena."
          },
          "status": {
            "$ref": "#/components/schemas/EvaluationStatus",
            "description": "Current evaluation status."
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if the evaluation failed."
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id",
            "description": "ID of the dataset being evaluated."
          },
          "judge_config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Config Id",
            "description": "ID of the judge configuration used for scoring."
          },
          "rubric_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric Id",
            "description": "ID of the rubric used for dimension-based scoring."
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config",
            "description": "Mode-specific configuration."
          },
          "result_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Count",
            "description": "Number of results (populated on detail endpoint)."
          },
          "average_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Score",
            "description": "Average score across results."
          },
          "pass_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Rate",
            "description": "Fraction of results that passed."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the evaluation was created."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Timestamp when the evaluation was last updated."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "mode",
          "status",
          "dataset_id",
          "judge_config_id",
          "config",
          "created_at",
          "updated_at"
        ],
        "title": "EvaluationResponse",
        "description": "Schema for an evaluation in API responses."
      },
      "EvaluationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "EvaluationStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_EvaluationResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvaluationResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "PaginatedResponse[EvaluationResponse]"
      },
...
    }
  }
}